pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "apscheduler>=3.11.0",
    "asgiref>=3.8.1",
    "beautifulsoup4>=4.13.3",
    "email-validator>=2.2.0",
    "fastapi>=0.115.11",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "httpx>=0.28.1",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.10.6",
    "uvicorn>=0.34.0",
]
